COMPLETE Phase 1: Extract massive drop handlers to useTimelineDropHandlers hook
✅ MAJOR MILESTONE: Complete hooks extraction (538 lines reduction)

Extracted:
- handleTrackDragOver (146 lines) - Overlap detection & drag preview logic
- handleTrackDrop (392 lines) - Timeline/text/media drop handling logic
- All state management and dependencies properly handled

Results:
- File size: 1,067 → 532 lines (50.1% reduction!)
- Added full implementations to useTimelineDropHandlers hook (644 lines)
- No TypeScript errors, clean compilation
- All drop functionality externalized to reusable hook

This completes the incomplete Phase 1 extraction and achieves the targeted
50%+ file size reduction through proper separation of concerns.
